[arg] Add new test witness of a current FP optn
When using 3rd order functions opt args in a 2nd order function passed
as argument are not marked as used although they are required by the
type of the 3rd order function.

- Test `examples/opt_args/opt_arg_in_arg_in_arg.ml` shows that the `?b`
opt arg of `g_skip` and `g_propagate` are both considered unused
although the 2 functions are passed as argument to `h_skip` and
`h_propagate`, which both expect their parameter to expect `?b`.

- Similarly, the `?a` of `g_skip`'s `f` is considered unused and,
although it appears unused because discarded in the body of `g_skip`,
`g_skip` is passed as argument to the `h_*` functions, which expect the
corresponding parameter to be a 2nd order function exepecting a function
expecting `?a`.
Changing the type of `g_skip` to `(unit -> 'a) -> ?b:'b -> unit -> 'a`
effectively fails to compile because of the missing `?a` in the first
parameter's type.