[src] compile in 5.3
Naively fix compilation issues. In particular, the use of
`cmt_declaration_dependecies` instead of `cmt_value_dependencies` is a
simple translation from the latter to the former and is not exactly
equivalent because `uid`s may not be found in the `cmt_uid_to_decl`
table. This is certainly the root cause of the introduced FP and FN