[src][input][1/n] rely on `.cmti` instead of `.cmi`
If a compilation has a`.mli`, then it has a `.cmti` which is more
complete than the `.cmi`. This extra information is useful to fill the
state's `file_infos`.
However, a `.ml` wihtout a corresponding `.mli` does not have a
corresponding `.cmti`. Hence, introducing even more mistakes for now.
Now that interfaces are read from `.cmti` instead of `.cmi`, the order
in which the files are read is reversed to ensure interfaces (`.cmti`)
are read before their corresponding implementation (`.cmt`)