[check][exp] fix expected results for `threshold-3-0.5`
In `threshold-3-0.5.exp`, there was only expected results for
`examples/matchopt.ml` and not for `examples/dir/matchopt.ml` in the
"optional arguments always/never used except once" sections.
Because these files are copies of each other and self-sustaining, the
expected results for both should be the same. Running on them separately
confirms the new expected results.
The current FN are newly discovered thanks to the reordering of files
to use `.cmti` instead of `.cmi` files but predate that change.