[src][state] expose `location_dependencies` through `state.file_infos`
Rather than  adding a `signature` field to the state to hold a
`uid_to_loc` table to then use in `DeadCode.load_file` to convert uids
into locations, to then convert `cmt_declaration_dependencies` into
a list of pairs of locations, directly construct that list and store
along with the `file_infos`. This move  is more consistent with the
nature, usage and provenance of the information : it is created using
info in `cmt_infos` and `cmti_infos` to replace the
`cmt_value_dependencies` field that existed prior to OCaml 5.3, and
should only live as long the corresponding `cmt_infos` live.